Fix compiler warnings from VS2019

+++ src/numerics/AdaptivePreconditioner.cpp
@@ -16,7 +16,7 @@ AdaptivePreconditioner::AdaptivePreconditioner()
 
 void AdaptivePreconditioner::setValue(size_t row, size_t col, double value)
 {
-    m_jac_trips.emplace_back(row, col, value);
+    m_jac_trips.emplace_back(static_cast<int>(row), static_cast<int>(col), value);
 }
 
 void AdaptivePreconditioner::stateAdjustment(vector_fp& state) {
@@ -47,7 +47,7 @@ void AdaptivePreconditioner::initialize(size_t networkSize)
         setIlutDropTol(1e-10);
     }
     if (m_drop_tol == 0) {
-        setIlutFillFactor(m_dim/4);
+        setIlutFillFactor(static_cast<int>(m_dim) / 4);
     }
     // update initialized status
     m_init = true;
